requirements.txt
PyQt5
pyqtgraph
pyserial
vtkplotlib
numpy-stl
pygame